% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{compute_c}
\alias{compute_c}
\title{ColumnStore version of \code{\link[dplyr:compute]{dplyr::compute()}}.}
\usage{
compute_c(sql, name, temporary, overwrite, ...)
}
\arguments{
\item{sql}{the sql to compute}

\item{name}{the name of the table to create (defaults to a new unique table name)}

\item{temporary}{whether to create a temporary table (defaults to TRUE if table name not specified, otherwise needs to be explicitly specified)}

\item{overwrite}{whether to overwrite existing tables (default to TRUE for temporary tables, FALSE otherwise)}

\item{...}{Other arguments passed on to \code{\link[dplyr:compute]{dplyr::compute()}},}
}
\value{
a dbplyr tbl referencing the table computed
}
\description{
ColumnStore version of \code{\link[dplyr:compute]{dplyr::compute()}}.
}
