% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{copy_to_c}
\alias{copy_to_c}
\title{Version of \code{\link[dplyr:copy_to]{dplyr::copy_to()}} that creates ColumnStore tables and has a better parameter order}
\usage{
copy_to_c(df, con, name, temporary, overwrite, ...)
}
\arguments{
\item{df}{the dataframe to copy to the SQL store}

\item{con}{the connection to the SQL store}

\item{name}{the name of the table to create (defaults to a new unique table name)}

\item{temporary}{whether to create a temporary table (defaults to TRUE if table name not specified, otherwise needs to be explicitly specified)}

\item{overwrite}{whether to overwrite existing tables (default to TRUE for temporary tables, FALSE otherwise)}

\item{...}{Other arguments passed on to \code{\link[dplyr:copy_to]{dplyr::copy_to()}},}
}
\value{
a dbplyr tbl referencing the table created
}
\description{
Version of \code{\link[dplyr:copy_to]{dplyr::copy_to()}} that creates ColumnStore tables and has a better parameter order
}
