% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{delete_table_finalizer}
\alias{delete_table_finalizer}
\title{Utility finalizer function to remove ColumnStore temporary tables}
\usage{
delete_table_finalizer(fe)
}
\arguments{
\item{fe}{finalizer environment containing connection (con) and table name (table_name)}
}
\description{
Utility finalizer function to remove ColumnStore temporary tables
}
\keyword{internal}
