% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{get_connection}
\alias{get_connection}
\title{Get a database connection as defined by a yaml configuration
or environment variables}
\usage{
get_connection(
  params = here("params.yaml"),
  secret = here("secret.yaml"),
  key = "db",
  bigint = "integer",
  ...
)
}
\arguments{
\item{params}{a params.yaml-file that defines some of db_host, db_name and db_user under a given key}

\item{secret}{a secret.yaml-file that defines some of db_host, db_name, db_user and db_pass under a given key}

\item{key}{the key in the yaml file to extract}

\item{bigint}{how should the connection convert bigints}

\item{...}{Other arguments passed on to \code{\link[DBI:dbConnect]{DBI::dbConnect()}},}
}
\value{
the MariaDB connection object
}
\description{
Get a database connection as defined by a yaml configuration
or environment variables
}
