% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil-package.R
\docType{package}
\name{hscidbutil-package}
\alias{hscidbutil}
\alias{hscidbutil-package}
\title{hscidbutil: HSCI Research Group Database Utilities}
\description{
Database utility functions used by the HSCI research group.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hsci-r.github.io/hscidbutil/}
}

}
\author{
\strong{Maintainer}: Eetu Mäkelä \email{eetu.makela@helsinki.fi} (\href{https://orcid.org/0000-0002-8366-8414}{ORCID})

}
\keyword{internal}
