% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{list_schemas}
\alias{list_schemas}
\title{List schemas in a database}
\usage{
list_schemas(con)
}
\arguments{
\item{con}{the connection to probe}
}
\value{
a list of the schemas defined in the database
}
\description{
List schemas in a database
}
