% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{register_tables}
\alias{register_tables}
\title{Register all tables in a schema as dbplyr tables in the given enviroment}
\usage{
register_tables(con, schemas, envir = .GlobalEnv)
}
\arguments{
\item{con}{the connection to probe}

\item{schemas}{the schemas to probe}

\item{envir}{the environment in which to register the tables}
}
\description{
Register all tables in a schema as dbplyr tables in the given enviroment
}
