% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{unique_table_name}
\alias{unique_table_name}
\title{Utility function to generate unique table names}
\usage{
unique_table_name()
}
\description{
Utility function to generate unique table names
}
\keyword{internal}
