% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{compute_a}
\alias{compute_a}
\title{Version of \code{\link[dplyr:compute]{dplyr::compute()}} that creates Aria tables.}
\usage{
compute_a(sql, name, temporary, overwrite, ...)
}
\arguments{
\item{sql}{the sql to compute}

\item{name}{the name of the table to create (defaults to a new unique table name)}

\item{temporary}{whether to create a temporary table (defaults to TRUE if table name not specified, otherwise needs to be explicitly specified)}

\item{overwrite}{whether to overwrite existing tables (default to TRUE for temporary tables, FALSE otherwise)}

\item{...}{Other arguments passed on to \code{\link[dplyr:compute]{dplyr::compute()}},}
}
\description{
Version of \code{\link[dplyr:compute]{dplyr::compute()}} that creates Aria tables.
}
