% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{compute_c}
\alias{compute_c}
\title{ColumnStore version of \code{\link[dplyr:compute]{dplyr::compute()}}.}
\usage{
compute_c(
  sql,
  name = unique_table_name(),
  overwrite = FALSE,
  temporary = TRUE,
  ...
)
}
\arguments{
\item{sql}{the sql to compute}

\item{name}{the name of the table to create (defaults to a new unique table name)}

\item{overwrite}{whether to overwrite existing tables (default \code{FALSE})}

\item{temporary}{whether to create a temporary table (default \code{TRUE})}

\item{...}{Other arguments passed on to \code{\link[dplyr:compute]{dplyr::compute()}},}
}
\description{
ColumnStore version of \code{\link[dplyr:compute]{dplyr::compute()}}.
}
