% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{get_connection}
\alias{get_connection}
\title{Get a database connection as defined by a yaml configuration
or environment variables}
\usage{
get_connection(
  db_params = here("db_params.yaml"),
  db_secret = here("db_secret.yaml"),
  bigint = "integer",
  ...
)
}
\arguments{
\item{db_params}{a db_params.yaml-file that defines db_host, db_name and db_user}

\item{db_secret}{a db_secret.yaml-file that defines db_pass and optionally overrides other values}

\item{bigint}{how should the connection convert bigints}

\item{...}{Other arguments passed on to \code{\link[DBI:dbConnect]{DBI::dbConnect()}},}
}
\description{
Get a database connection as defined by a yaml configuration
or environment variables
}
