% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscidbutil.R
\name{list_temporary_tables}
\alias{list_temporary_tables}
\title{List all "temporary" tables starting with \code{tmp_} in the given schema}
\usage{
list_temporary_tables(con, schemas)
}
\arguments{
\item{con}{the connection to probe}

\item{schemas}{the schemas to probe}
}
\description{
List all "temporary" tables starting with \code{tmp_} in the given schema
}
